% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analyseCosts}
\alias{analyseCosts}
\title{Returns cost-effectiveness for a single birthcohort in a single country}
\usage{
analyseCosts(results, vaccine_cost, gdp_per_capita)
}
\arguments{
\item{results}{Data.table (required): results from RunCohort()}

\item{vaccine_cost}{Number (required): cost of a single vaccine}

\item{gdp_per_capita}{Number (required): GDP per capita}
}
\value{
Data.table with cost-analysis
}
\description{
Usually called using RunCountry(..., analyseCosts=TRUE)
}
\examples{
analyseCosts(RunCountry("AFG"), 100, 561)
}
