% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_functions.r
\name{BatchRun}
\alias{BatchRun}
\title{BatchRun}
\usage{
BatchRun(countries = -1, coverage = -1, agevac = -1,
  agecohort = -1, canc.inc = "2012", daly.canc.diag = 0.08,
  daly.canc.terminal = 0.78, sens = -1, unwpp_mortality = FALSE,
  year_born = -1, year_vac = -1, runs = 1,
  vaccine_efficacy_beforesexdebut = 1,
  vaccine_efficacy_aftersexdebut = 0, log = -1,
  by_calendaryear = FALSE, use_proportions = TRUE,
  analyseCosts = FALSE, psa = 0, psa_vals = ".data.batch.psa")
}
\arguments{
\item{use_proportions}{}
}
\value{
#
}
\description{
BatchRun
}
\examples{
#
}
