% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.R
\name{calcAnnoCTSS}
\alias{calcAnnoCTSS}
\title{Calculate the total tags for each annotation type.}
\usage{
calcAnnoCTSS(data, txModels, uniqueCTSS = FALSE)
}
\arguments{
\item{data}{A \code{RangedSummarizedExperiment} object containing the CTSSs.}

\item{txModels}{A \code{TxDb} object containing the the transcript models for annotation.}

\item{uniqueCTSS}{A \code{logical} indicating whether to count only unique CTSSs (TRUE) or to sum the total tags (FALSE).}
}
\value{
A data frame containing the total tags for each annotation type.
}
\description{
Calculate the total tags for each annotation type.
}
