% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{calcBatchSupport}
\alias{calcBatchSupport}
\title{Calculate batch support for expression data.}
\usage{
calcBatchSupport(
  object,
  batch,
  inputAssay = "counts",
  outputColumn = "batchSupport",
  unexpressed = 0,
  minSamples = 1
)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object.}

\item{batch}{A \code{factor} indicating the batch of each sample.}

\item{inputAssay}{The assay to use.}

\item{outputColumn}{The name of the column to store the batch support.}

\item{unexpressed}{The threshold below which a unit (CTSS, gene, etc.) 
is considered unexpressed.}

\item{minSamples}{The minimum number of samples in which a unit must be
expressed to be considered supported.}
}
\value{
A \code{SummarizedExperiment} object with the batch support column.
}
\description{
Calculate batch support for expression data.
}
