% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{calcCTSSComplexity}
\alias{calcCTSSComplexity}
\title{Calculate the CTSS complexity of a CTSS dataset.}
\usage{
calcCTSSComplexity(
  object,
  step = 1e+06,
  CTSSunexpressed = 1,
  minCTSSsupport = 2
)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} resulting from
CAGEfightR::quantifyCTSSs().}

\item{step}{The step size for the subsampling.}

\item{CTSSunexpressed}{The count threshold for considering a CTSS to be
unexpressed.}

\item{minCTSSsupport}{The minimum number of samples a CTSS must be
expressed in to be considered.}
}
\value{
A data frame containing the number of CTSSs (column "numberCTSSs")
for each sample and target (column "totalTags").
}
\description{
This function calculates the CTSS complexity of a given CTSS
dataset.
}
