% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{calcDivergentLociComplexity}
\alias{calcDivergentLociComplexity}
\title{Calculate the divergent loci complexity of a CTSS dataset.}
\usage{
calcDivergentLociComplexity(
  object,
  loci,
  step = 1e+06,
  CTSSunexpressed = 1,
  lociunexpressed = 2,
  requirebidirectional = FALSE,
  minCTSSsupport = 2
)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} resulting from
CAGEfightR::quantifyCTSSs().}

\item{loci}{A \code{GRanges} object containing the tag clusters from which 
divergent loci should be called.}

\item{step}{The step size for the subsampling.}

\item{CTSSunexpressed}{The count threshold for considering a CTSS to be
unexpressed.}

\item{lociunexpressed}{The count threshold for considering a tag cluster
to be unexpressed.}

\item{requirebidirectional}{Logical indicating whether to require 
bidirectional transcription for a divergent locus from the same sample.}

\item{minCTSSsupport}{The minimum number of samples a CTSS must be
expressed in to be considered.}
}
\value{
A data frame containing the number of divergent loci (column 
"numberLoci") for each sample and target (column "totalTags").
}
\description{
This function calculates the divergent loci complexity of a given CTSS
dataset.
}
