% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{calcGeneComplexity}
\alias{calcGeneComplexity}
\title{Calculate the gene complexity of a CTSS dataset.}
\usage{
calcGeneComplexity(
  object,
  txModels,
  step = 1e+06,
  CTSSunexpressed = 1,
  geneunexpressed = 9,
  minCTSSsupport = 2
)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} resulting from
\code{CAGEfightR::quantifyCTSSs()}.}

\item{txModels}{A \code{TxDb} object containing the gene models.}

\item{step}{The step size for the subsampling.}

\item{CTSSunexpressed}{The count threshold for considering a CTSS to be
unexpressed.}

\item{geneunexpressed}{The count threshold for considering a gene to be
unexpressed.}

\item{minCTSSsupport}{The minimum number of samples a CTSS must be
expressed in to be considered.}
}
\value{
A data frame containing the number of genes (column "numberGenes") 
for each sample and target (column "totalTags").
}
\description{
This function calculates the gene complexity of a given CTSS
dataset.
}
