% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{cast_rse_to_granges}
\alias{cast_rse_to_granges}
\title{Convert SummarizedExperiment to GRanges with Assay Data}
\usage{
cast_rse_to_granges(rse, assay = "counts", coln_assay = 1, colname = "score")
}
\description{
This function converts a `SummarizedExperiment` object to a `GRanges` object
and adds a specified assay data column to the resulting `GRanges` object.
}
