% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{check_python_dependencies}
\alias{check_python_dependencies}
\title{Internal utility to check that all required Python packages are installed
and meet the minimum version requirements. Raises an error if any package
is missing or incompatible. Not exported.}
\usage{
check_python_dependencies(required_packages)
}
\description{
Internal utility to check that all required Python packages are installed
and meet the minimum version requirements. Raises an error if any package
is missing or incompatible. Not exported.
}
