% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{check_valid_profile_rownames}
\alias{check_valid_profile_rownames}
\title{Check presence and format of rownames in a profile matrix}
\usage{
check_valid_profile_rownames(profile_matrix, chr_name)
}
\description{
Check presence and format of rownames in a profile matrix
}
