% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkSize}
\alias{checkSize}
\title{Checks whether the size of a variable is larger than 0}
\usage{
checkSize(v)
}
\arguments{
\item{v}{Variable (required)}
}
\value{
Logical: TRUE if size is not 0, false if size is 0
}
\description{
Used to determine that all required variables are passed to a function
Checks whether a vector has length > 0 or a data.table/data.frame has nrow > 0
}
\examples{
x <- c()
checkSize(x)

x <- c(2,5)
checkSize(x)

A <- c()
B <- c(1,2,3)
sapply(c("A","B"),function(x){checkSize(get(x))})
}
