% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{choose_required_packages}
\alias{choose_required_packages}
\title{Dynamically choose required packages based on numpy version}
\usage{
choose_required_packages(required_numpy1, required_numpy2)
}
\description{
Dynamically choose required packages based on numpy version
}
