% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{conditionalNormalize}
\alias{conditionalNormalize}
\title{Conditional normalization method (e.g. on GC content) based on approach 
described in Pickrell, et al, 2010, Nature.}
\usage{
conditionalNormalize(
  object,
  inputAssay = "counts",
  outputAssay = "normalized",
  conditionalColumn = "GC",
  offsetAssay = NULL,
  bins = 200,
  sizeFactors = NULL,
  minCount = 1,
  minSamples = 1,
  aggregate.fn = sum
)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object.}

\item{inputAssay}{The name of the assay to use for the calculation.}

\item{outputAssay}{The name of the assay to use for the output.}

\item{conditionalColumn}{The column in the metadata to use for the 
conditional normalization.}

\item{offsetAssay}{The name of the assay to use for the offset.}

\item{bins}{The number of bins to use for the conditional normalization.}

\item{sizeFactors}{The size factors to use for the normalization.}

\item{minCount}{The minimum count to consider a unit (CTSS, gene, etc.) 
as expressed.}

\item{minSamples}{The minimum number of samples to consider a unit 
as expressed.}

\item{aggregate.fn}{The function to use for aggregation.}
}
\value{
A \code{SummarizedExperiment} object with the normalized counts.
}
\description{
Conditional normalization method (e.g. on GC content) based on approach 
described in Pickrell, et al, 2010, Nature.
}
