% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{convert_rowname_to_nostrand}
\alias{convert_rowname_to_nostrand}
\title{Convert strand suffix in row names to "*"}
\usage{
convert_rowname_to_nostrand(strand_str)
}
\description{
Convert strand suffix in row names to "*"
}
