% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{corr_decompose}
\alias{corr_decompose}
\title{Decompose tag cluster into subclusters according to CTSS expression
correlation across samples.}
\usage{
corr_decompose(
  rse,
  gr,
  assay = "TPM",
  thres = 0.25,
  merge = TRUE,
  scale = TRUE
)
}
\arguments{
\item{rse}{A \code{RangedSummarizedExperiment} object containing the CTSSs.}

\item{gr}{A \code{GRanges} object of tag clusters.}

\item{assay}{The assay to use for the correlation calculation.}

\item{thres}{The threshold for the Bayesian change point analysis.}

\item{merge}{If TRUE, subclusters will be merged if positively correlated.}

\item{scale}{If TRUE, the assay values will be scaled before correlation 
calculation.}
}
\value{
An \code{IRanges} object of decomposed tag clusters.
}
\description{
Decompose tag cluster into subclusters according to CTSS expression
correlation across samples.
}
