% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{corr_decompose_changepoint}
\alias{corr_decompose_changepoint}
\title{Decompose tag cluster into subclusters according to CTSS expression
correlation across samples using changepoint detection.}
\usage{
corr_decompose_changepoint(
  rse,
  gr,
  assay = "TPM",
  penalty = "MBIC",
  pen.value = 0,
  method = "PELT",
  minseglen = 2,
  merge = TRUE,
  scale = TRUE,
  ...
)
}
\arguments{
\item{rse}{A \code{RangedSummarizedExperiment} object containing the CTSSs.}

\item{gr}{A \code{GRanges} object of tag clusters.}

\item{assay}{The assay to use for the correlation calculation.}

\item{penalty}{Penalty used by \code{changepoint::cpt.mean()}.}

\item{pen.value}{Numeric penalty value used when \code{penalty = "Manual"}.}

\item{method}{Changepoint search method used by \code{changepoint::cpt.mean()}.}

\item{minseglen}{Minimum segment length passed to \code{changepoint::cpt.mean()}.}

\item{merge}{If TRUE, subclusters will be merged if positively correlated.}

\item{scale}{If TRUE, the assay values will be scaled before correlation
calculation.}

\item{...}{Additional arguments passed to \code{changepoint::cpt.mean()}.}
}
\value{
An \code{IRanges} object of decomposed tag clusters.
}
\description{
Decompose tag cluster into subclusters according to CTSS expression
correlation across samples using changepoint detection.
}
