% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{corrProfiles}
\alias{corrProfiles}
\title{Calculate the auto correlation of CTSSs over tag clusters according to 
pooled values and the cross correlation with opposite strand CTSSs}
\usage{
corrProfiles(object, pooled, flank = 300, lag.max = 300)
}
\arguments{
\item{object}{A \code{GRanges} object of tag clusters.}

\item{pooled}{A \code{RangedSummarizedExperiment} object containing the CTSSs and
the pooled values in the score column (resulting from 
\code{CAGEfightR::calcPooled()}).}

\item{flank}{The number of base pairs to consider around each tag cluster.}

\item{lag.max}{The maximum lag to consider for the correlation.}
}
\value{
A \code{list} containing the auto correlation matrix and cross correlation
matrix
}
\description{
Calculate the auto correlation of CTSSs over tag clusters according to 
pooled values and the cross correlation with opposite strand CTSSs
}
