% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{create_granges_from_bed}
\alias{create_granges_from_bed}
\title{Create a GRanges object from a BED data.table}
\usage{
create_granges_from_bed(bed_file)
}
\arguments{
\item{bed_file}{A `data.table` containing the BED file data.}
}
\value{
A `GRanges` object.
}
\description{
This function converts a `data.table`
containing BED file data into a `GRanges` object.
The required columns are extracted and
used to define the `GRanges` object, and the remaining
columns are added as metadata.
}
