% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{create_granges_from_rownames}
\alias{create_granges_from_rownames}
\title{Create GRanges object from row names like "chr:start-end;strand"}
\usage{
create_granges_from_rownames(row_names_str)
}
\description{
Create GRanges object from row names like "chr:start-end;strand"
}
