% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CreatePsaData}
\alias{CreatePsaData}
\title{Generate Latin hyper cube sample of parameters for sensitivity analysis}
\usage{
CreatePsaData(
  country_codes,
  vaccine = "4vHPV",
  psa_runs = 0,
  seed_state = 1,
  psadat_file = "psadat.csv",
  psadat_vimc_file = "psadat_vimc.csv"
)
}
\arguments{
\item{country_codes}{ISO3 country codes of countries}

\item{vaccine}{bivalent/quadrivalent or nonavalent HPV vaccine}

\item{psa_runs}{integer, simulation runs for sensitivity analysis}

\item{seed_state}{integer, seed value for random number generator}

\item{psadat_file}{character string, file to save Latin hyper cube
sample of input parameters}

\item{psadat_vimc_file}{character string, file to save Latin hyper cube
sample of input parameters (VIMC format)}
}
\value{
Null return value; disease burden estimates are saved to corresponding files
}
\description{
Generate Latin hyper cube sample of input parameters based on their
  distributions for probabilistic sensitivity analysis.
}
\examples{
  # CreatePsaDataNew (
  #   country_codes    = c("AFG", "ALB"),
  #   vaccine          = "4vHPV",
  #   psa_runs         = 200,
  #   seed_state       = 1,
  #   psadat_file      = "psadat.csv",
  #   psadat_vimc_file = "psadat_vimc.csv")
}
