% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{cumulativeFractionsAroundLoci}
\alias{cumulativeFractionsAroundLoci}
\title{Calculate the cumulative fraction of CTSSs around loci, considering unique
tags and the fraction of loci with CTSSs.}
\usage{
cumulativeFractionsAroundLoci(
  loci,
  ctss,
  max_dist = 1000,
  inputAssay = "counts"
)
}
\arguments{
\item{loci}{A \code{GRanges} object of loci to focus on.}

\item{ctss}{A \code{RangedSummarizedExperiment} object containing the CTSSs.}

\item{max_dist}{The number of base pairs to consider around each locus.}

\item{inputAssay}{The assay in the \code{ctss} object to use for counting tags (default is "counts").}
}
\value{
A list of three data frames: cumulative fraction of tags, cumulative fraction of unique tags, and cumulative fraction of loci with CTSSs, for each sample.
}
\description{
Calculate the cumulative fraction of CTSSs around loci, considering unique
tags and the fraction of loci with CTSSs.
}
