% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.cecx_5y_prevalence}
\alias{data.cecx_5y_prevalence}
\title{5-year prevalence of cervical cancer}
\format{A data table with 185 observations of 103 variables.
\describe{
  \item{Country}{Country name}
  \item{iso3}{ISO3 country code}
  \item{Source}{Data source}
  \item{Year}{Data source (year)}
  \item{0..100}{Age 0-100}
}}
\source{
\url{https://gco.iarc.fr/today/online-analysis-table}
}
\usage{
data.cecx_5y_prevalence
}
\description{
A dataset containing the 5-year prevalence (proportion) of cervical cancer in 185 countries, as reported by
    IARC's Globocan 2018 database.
}
\details{
As per IARC definition -- The (5-year) prevalence of a given cancer
    is the number of individuals within a defined population who have been diagnosed with that
    cancer (within 5 years) and who are still alive at a given point in time (i.e. the survivors).
}
\keyword{datasets}
