% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.costcecx}
\alias{data.costcecx}
\title{Cost of cervical cancer treatment}
\format{
A data table with 194 observations of 7 variables.
\describe{
  \item{country}{Country name}
  \item{cancer_cost}{cost per cancer episode, in 2017 US$}
  \item{cancer_cost_adj}{cost per cancer episode -- adjusted, in 2017 US$ --
        “adjusted” cancer costs are based on a GDP/capita based adjustment
        within the region}
  \item{iso3}{ISO3 country code}
  \item{cancer_cost_2011}{cost per cancer episode, in 2011 US$}
  \item{cancer_cost_adj_2011}{cost per cancer episode --
        adjusted, in 2011 US$ -- “adjusted” cancer costs are based on a
        GDP/capita based adjustment within the region}
  \item{inflation_factor}{Inflation factor from 2011 to 2017 estimated from
        Inflation, GDP deflator (annual \%) --
        https://data.worldbank.org/indicator/NY.GDP.DEFL.KD.ZG}
}
}
\usage{
data.costcecx
}
\description{
A dataset containing the cost of cervical cancer treatment.
}
\keyword{datasets}
