% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.countryname}
\alias{data.countryname}
\title{Country names and codes.}
\format{A data table with 251 observations of 13 variables (251 row * 13 col).
\describe{
  \item{name1, name2, name3, name4}{country names}
  \item{name2}{}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
data.countryname
}
\description{
A dataset containing the country names and codes (ISO/WHO/WB/UN).
}
\keyword{datasets}
