% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.cross_protection}
\alias{data.cross_protection}
\title{Cross-protection effect of bivalent and quadrivalent vaccines}
\format{
A data table with 5 observations of 7 variables.
\describe{
  \item{hpv_type}{HPV strain type}
  \item{bivalent}{Cross-protection of bivalent vaccine against the corresponding HPV strain type (mean)}
  \item{bivalent_low}{Cross-protection of bivalent vaccine against the corresponding HPV strain type (lower bound of 95\% uncertainty interval)}
  \item{bivalent_high}{Cross-protection of bivalent vaccine against the corresponding HPV strain type (upper bound of 95\% uncertainty interval)}
  \item{quadrivalent}{Cross-protection of quadrivalent vaccine against the corresponding HPV strain type (mean)}
  \item{quadrivalent_low}{Cross-protection of quadrivalent vaccine against the corresponding HPV strain type (lower bound of 95\% uncertainty interval)}
  \item{quadrivalent_high}{Cross-protection of quadrivalent vaccine against the corresponding HPV strain type (upper bound of 95\% uncertainty interval)}
}
}
\source{
{Malagón T, Drolet M, Boily MC, Franco EL, Jit M, Brisson J, Brisson M.
Cross-protective efficacy of two human papillomavirus vaccines: a systematic review and meta-analysis.
Lancet Infect Dis. 2012 Oct;12(10):781-9. \url{https://doi.org/10.1016/S1473-3099(12)70187-1}}
}
\usage{
data.cross_protection
}
\description{
A dataset containing cross-protection effect of bivalent and quadrivalent vaccines
}
\keyword{datasets}
