% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.mortall.unwpp.nqx}
\alias{data.mortall.unwpp.nqx}
\title{UNWPP life table}
\format{A data table with 122850 observations of 8 variables.
\describe{
  \item{country_code_numeric}{Country code numeric}
  \item{country_code}{ISO3 country code}
  \item{country}{Country name}
  \item{age_from}{age from (start-age)}
  \item{age_to}{age to (end-age)}
  \item{year}{Year}
  \item{gender}{Gender}
  \item{value}{nqx - probability of dying between ages x and x+n}
}}
\source{
\url{https://population.un.org/wpp/}
}
\usage{
data.mortall.unwpp.nqx
}
\description{
A dataset containing the UNWPP life table (World Population Prospects 2019).
}
\keyword{datasets}
