% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.mortcecx}
\alias{data.mortcecx}
\title{Mortality from cervical cancer, by age and country}
\format{A data table with 185 observations of 103 variables.
\describe{
  \item{Country}{Country name}
  \item{0..100}{Age 0-100}
  \item{iso3}{ISO3 country code}
}}
\source{
\url{https://gco.iarc.fr/today/online-analysis-table}
}
\usage{
data.mortcecx
}
\description{
A dataset containing the mortality from cervical cancer in 185 countries, as reported by
    IARC's Globocan 2018 database.
}
\keyword{datasets}
