% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.mortcecx2020}
\alias{data.mortcecx2020}
\title{Mortality from cervical cancer, by age and country}
\format{
A data table with 185 observations of 103 variables.
\describe{
  \item{Country}{Country name}
  \item{iso3}{ISO3 country code}
  \item{Source}{Data source}
  \item{Year}{Data source (year)}
  \item{0..100}{Age 0-100 - Crude rates, cervix uteri, females, by age; annual rate per individual}
}
}
\source{
\url{https://gco.iarc.fr/today/online-analysis-table?v=2020}
}
\usage{
data.mortcecx2020
}
\description{
A dataset containing the mortality (crude rates)
    from cervical cancer in 185 countries,
    as reported by IARC's Globocan 2020 database.
}
\keyword{datasets}
