% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.mortcecx_ui}
\alias{data.mortcecx_ui}
\title{Mortality from cervical cancer with uncertainty intervals, all ages and by country}
\format{A data table with 185 observations of 5 variables.
\describe{
  \item{Country}{Country name}
  \item{iso3}{ISO3 country code}
  \item{Mid}{Estimated number of deaths in 2018, cervix uteri, females, all ages (mean)}
  \item{Low}{Estimated number of deaths in 2018, cervix uteri, females, all ages (lower bound of 95\% uncertainty interval)}
  \item{High}{Estimated number of deaths in 2018, cervix uteri, females, all ages (upper bound of 95\% uncertainty interval)}
}}
\source{
\url{https://gco.iarc.fr/today/online-analysis-table}
}
\usage{
data.mortcecx_ui
}
\description{
A dataset containing the number of deaths from cervical cancer with uncertainty
    intervals in 185 countries, as reported by IARC's Globocan 2018 database.
    Estimated number of deaths in 2018, cervix uteri, females, all ages.
}
\keyword{datasets}
