% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.pop}
\alias{data.pop}
\title{UNWPP population estimates}
\format{
A data table with 3392220 observations of 8 variables.
\describe{
  \item{country_code_numeric}{Country code numeric}
  \item{country_code}{ISO3 country code}
  \item{country}{Country name}
  \item{age_from}{age from (start-age)}
  \item{age_to}{age to (end-age)}
  \item{year}{Year}
  \item{gender}{Gender}
  \item{value}{Population size}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Population/}

{VIMC}
}
\usage{
data.pop
}
\description{
A dataset containing the UNWPP population estimates -- World Population Prospects 2019.
}
\keyword{datasets}
