% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.popproj}
\alias{data.popproj}
\title{Population projections of 5-year old girls}
\format{A data table with 98 observations of 91 variables.
\describe{
  \item{iso3}{ISO3 country code}
  \item{2011..2100}{Year -- 2011..2100}
}}
\source{
\url{}
}
\usage{
data.popproj
}
\description{
A dataset containing population projections of 5-year old girls.
}
\keyword{datasets}
