% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.quality}
\alias{data.quality}
\title{Data quality of incidence and mortality}
\format{A data table with 186 observations of 4 variables.
\describe{
  \item{Country}{Country name}
  \item{Incidence}{Quality of cervical cancer incidence data}
  \item{Mortality}{Quality of cervical cancer mortality data}
  \item{iso3}{ISO3 country code}
}}
\source{
\url{}
}
\usage{
data.quality
}
\description{
A dataset indicating data quality of cervical cancer incidence and mortality.
}
\keyword{datasets}
