% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.sexual_debut}
\alias{data.sexual_debut}
\title{Sexual debut data}
\format{A data table with 94 observations of 14 variables.
\describe{
  \item{V1}{Row number}
  \item{iso2}{ISO2 country code}
  \item{country}{Country name}
  \item{iso3}{ISO3 country code}
  \item{who}{WHO region}
  \item{X15}{?}
  \item{X18}{?}
  \item{X20}{?}
  \item{X22}{?}
  \item{X25}{?}
  \item{Never}{Proportion of people who had not sexually debuted}
  \item{cluster.id}{?}
  \item{a}{Parameter for sexual debut curve (logistic model)}
  \item{b}{Parameter for sexual debut curve (logistic model)}
}}
\source{
\url{}
}
\usage{
data.sexual_debut
}
\description{
A dataset containing sexual debut data and (2) parameters for the sexual debut curve (logistic model).
}
\keyword{datasets}
