% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Decompose tag clusters according to pooled values and a decomposition 
function (implemented functions: summit_decompose and local_maxima_decompose)}
\usage{
decompose(object, pooled, fn = summit_decompose, ...)
}
\arguments{
\item{object}{A \code{GRanges} object of tag clusters.}

\item{pooled}{A \code{RangedSummarizedExperiment} object containing the 
CTSSs.}

\item{fn}{A function that carries out the actual decomposition. 
Implemented functions are summit_decompose and local_maxima_decompose.
It takes as input a Views object (and optional arguments) and returns an
IRanges object}

\item{...}{Additional arguments passed on to the decomposition function.}
}
\value{
A \code{GRanges} object of decomposed tag clusters.
}
\description{
Decompose tag clusters according to pooled values and a decomposition 
function (implemented functions: summit_decompose and local_maxima_decompose)
}
