% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decomposeCorr}
\alias{decomposeCorr}
\title{Decompose tag clusters according to CTSS correlations across samples}
\usage{
decomposeCorr(object, ctss, fn = corr_decompose, ...)
}
\arguments{
\item{object}{A \code{GRanges} object of tag clusters.}

\item{ctss}{A \code{RangedSummarizedExperiment} object containing the CTSSs.}

\item{fn}{A function that carries out the actual decomposition. 
Implemented functions are: corr_decompose and corr_decompose_changepoint.
It takes as input a RangedSummarizedExperiment object (and optional 
arguments) and returns an IRanges object.}

\item{...}{Additional arguments passed on to the decomposition function.}
}
\value{
A \code{GRanges} object of decomposed tag clusters.
}
\description{
Decompose tag clusters according to CTSS correlations across samples
}
