% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergent.R
\name{divergentLoci}
\alias{divergentLoci}
\title{Identify divergent loci from tag clusters.}
\usage{
divergentLoci(
  object,
  ctss,
  max_gap = 400,
  win_size = 200,
  inputAssay = "counts"
)
}
\arguments{
\item{object}{A \code{GRanges} object containing the tag clusters.}

\item{ctss}{A \code{RangedSummarizedExperiment} resulting from
\code{CAGEfightR::quantifyCTSSs()}.}

\item{max_gap}{Maximum distance between tag clusters to be considered
part of a divergent locus.}

\item{win_size}{Size of the flanking windows around the midpoint of the
divergent locus.}

\item{inputAssay}{The assay to use for quantification.}
}
\value{
A \code{GRanges} object containing the divergent loci.
}
\description{
Identify divergent loci from tag clusters.
}
