% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergent.R
\name{divergentLociTCsSummit}
\alias{divergentLociTCsSummit}
\title{Identify divergent loci from tag clusters with summit centering. The function is similar to \code{divergentLociSummit} 
but does not require tag clusters to be passed to the function. This is handeled internally by calling 
\code{CAGEfightR::clusterUnidirectionally} on the pooled CTSS signal.}
\usage{
divergentLociTCsSummit(
  ctss,
  max_gap = 400,
  win_size = 200,
  inputAssay = "counts",
  callingAssay = "counts"
)
}
\arguments{
\item{ctss}{A \code{RangedSummarizedExperiment} resulting from
\code{CAGEfightR::quantifyCTSSs()}.}

\item{max_gap}{Maximum distance between tag clusters to be considered
part of a divergent locus.}

\item{win_size}{Size of the flanking windows around the midpoint of the
divergent locus.}

\item{inputAssay}{The assay to use for quantification.}

\item{callingAssay}{The assay to use for calling TCs.}
}
\value{
A \code{GRanges} object containing the divergent loci.
}
\description{
Identify divergent loci from tag clusters with summit centering. The function is similar to \code{divergentLociSummit} 
but does not require tag clusters to be passed to the function. This is handeled internally by calling 
\code{CAGEfightR::clusterUnidirectionally} on the pooled CTSS signal.
}
