% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{DMadjustedCV}
\alias{DMadjustedCV}
\title{Calculate the distance to the median (DM) adjusted CV2 for a given 
\code{RangedSummarizedExperiment} object. Inspired by the method outlined in
Kolodziejczyk et al., 2015, Cell Stem Cell.}
\usage{
DMadjustedCV(object, inputAssay = "TPM", prefix = "")
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} object.}

\item{inputAssay}{The name of the assay to use for the calculation.}

\item{prefix}{The prefix to use for the output columns.}
}
\value{
A \code{RangedSummarizedExperiment} object with the adjusted CV2
in the metadata columns.
}
\description{
Calculate the distance to the median (DM) adjusted CV2 for a given 
\code{RangedSummarizedExperiment} object. Inspired by the method outlined in
Kolodziejczyk et al., 2015, Cell Stem Cell.
}
