% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{EmulateVaccineImpactVimcStochastic}
\alias{EmulateVaccineImpactVimcStochastic}
\title{Emulate vaccine impact estimates (VIMC stochastic runs)}
\usage{
EmulateVaccineImpactVimcStochastic(disease_burden_template_file,
  centralBurdenResultsFile, psaData, diseaseBurdenStochasticFolder,
  diseaseBurdenStochasticFile, psa_runs, countryCodes = -1,
  vaccination_scenario)
}
\arguments{
\item{disease_burden_template_file}{csv file (required),
central disease burden template (vimc format); add column with run_id to get
stochastic disease burden template}

\item{centralBurdenResultsFile}{csv file (required), central disease burden
estimates pre and post vaccination}

\item{psaData}{data table (required), latin hyper cube sample of input parameters}

\item{diseaseBurdenStochasticFolder}{character string (required),
stochastic disease burden estimates folder (output)}

\item{diseaseBurdenStochasticFile}{character string (required),
stochastic disease burden estimates file(s) (output)}

\item{psa_runs}{integer (required), simulation runs for sensitivity analysis}

\item{countryCodes}{list (optional), If country codes are provided,
stochastic burden estimates are generated for these countries.
If set to -1, then stochastic burden estimates are generated for the
countries included in the central burden estimates.}

\item{vaccination_scenario}{logical (required), generate stochastic burden
estimates for (vaccination) or (no vaccination) scenario}
}
\value{
Null return value; disease burden estimates are saved to corresponding files
}
\description{
Emulate vaccine impact estimates for VIMC stochastic runs/sensitivity
  analysis. The inputs are central disease burden estimates, input
  parameter distributions (latin hyper sampling), runs for sensitivity analysis,
  and filename for stochastic burden estimates. The outputs are stochastic
  disease burden estimates (full results file plus a file per country).
}
\details{
Stochastic disease burden estimates are generated.
    (i)  full results files
           1 full results file for pre-vaccination (optional)
           1 full results file for post-vaccination
    (ii) 1 file per country for all runs
           1 file per country for all runs -- pre-vaccination (optional)
           1 file per country for all runs -- post-vaccination
}
