% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Estimate_all_cecx_burden_central}
\alias{Estimate_all_cecx_burden_central}
\title{Generate vaccine impact estimates - HPV all types (VIMC central run)}
\usage{
Estimate_all_cecx_burden_central(
  cecx_burden_file,
  cecx_burden_file_all,
  vaccine,
  disease_burden_template_file = "",
  disease_burden_no_vaccination_file_all = "",
  disease_burden_vaccination_file_all = "",
  vimc_output = TRUE
)
}
\arguments{
\item{cecx_burden_file}{csv file (input), disease burden estimates
(cecx burden attributable to HPV 16/18)}

\item{cecx_burden_file_all}{csv file (output), disease burden estimates
(cecx burden attributable to HPV all types)}

\item{vaccine}{character, bivalent/quadrivalent/nonovalent HPV vaccine}

\item{disease_burden_template_file}{csv file (input), disease burden template
(vimc format)}

\item{disease_burden_no_vaccination_file_all}{csv file (output), disease burden estimates
for no vaccination (cecx burden attributable to HPV all types - vimc format)}

\item{disease_burden_vaccination_file_all}{csv file (output), disease burden estimates
for vaccination (cecx burden attributable to HPV all types - vimc format)}
}
\value{
Null return value; disease burden estimates are saved to corresponding files
}
\description{
Generate vaccine impact estimates (HPV all types) for VIMC central runs.
  The inputs are cecx burden estimates pre- and post-vaccination with
    4vHPV / 9vHPV at different ages.
  The outputs are cecx burden estimates pre- and post-vaccination with
    HPV all types at different ages.
}
\details{
Three disease burden estimates (HPV all types) are generated.
  (i)   disease burden estimates for no vaccination (vimc format)
  (ii)  disease burden estimates for vaccination (vimc format)
  (iii) disease burden estimates for vaccination (pre- and post-vaccination)
          and includes YLDs and YLLs
}
\examples{
    Estimate_all_cecx_burden_central (
      cecx_burden_file                       = "central_burden_results_file.csv",
      cecx_burden_file_all                   = "central_burden_results_file_all.csv",
      vaccine                                = "4vHPV",
      disease_burden_template_file           = "central_burden_template_file.csv",
      disease_burden_no_vaccination_file_all = "central_burden_no_vaccination_file_all_vimc_format.csv",
      disease_burden_vaccination_file_all    = "central_burden_vaccination_file_all_vimc_format.csv",
      vimc_output                            = TRUE)

}
