% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{EstimateVaccineImpactVimcCentral}
\alias{EstimateVaccineImpactVimcCentral}
\title{Generate vaccine impact estimates (VIMC central run)}
\usage{
EstimateVaccineImpactVimcCentral(vaccine_coverage_file,
  disease_burden_template_file, disease_burden_no_vaccination_file,
  disease_burden_vaccination_file, disease_burden_results_file,
  campaign_vaccination, routine_vaccination, vaccine = "4vHPV")
}
\arguments{
\item{vaccine_coverage_file}{csv file (input), vaccine coverage data
(vimc format)}

\item{disease_burden_template_file}{csv file (input), disease burden template
(vimc format)}

\item{disease_burden_no_vaccination_file}{csv file (output), disease burden estimates
for no vaccination (vimc format)}

\item{disease_burden_vaccination_file}{csv file (output), disease burden estimates
for vaccination (vimc format)}

\item{disease_burden_results_file}{csv file (output), disease burden estimates
pre-vaccination and post-vaccination}

\item{campaign_vaccination}{logical, indicates campaign vaccination}

\item{routine_vaccination}{logical, indicates routine vaccination}
}
\value{
Null return value; disease burden estimates are saved to corresponding files
}
\description{
Generate vaccine impact estimates for VIMC central runs. The inputs are
  vaccine coverage and disease burden template files and outputs are
  disease burden estimates (pre-vaccination and post-vaccination).
}
\details{
Three disease burden estimates are generated.
  (i)   disease burden estimates for no vaccination (vimc format)
  (ii)  disease burden estimates for vaccination (vimc format)
  (iii) disease burden estimates for vaccination (pre- and post-vaccination)
  and includes YLDs and YLLs
}
\examples{
  # EstimateVaccineImpactVimcCentral (
  #   vaccine_coverage_file              = "coverage_hpv-routine-default.csv",
  #   disease_burden_template_file       = "central-burden-template.csv",
  #   disease_burden_no_vaccination_file = "central_burden_no_vaccination.csv",
  #   disease_burden_vaccination_file    = "central_burden_vaccination.csv",
  #   disease_burden_results_file        = "central_burden_results.csv",
  #   routine_vaccination                = TRUE,
  #   campaign_vaccination               = TRUE)
}
