% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIME.R
\name{EucDist}
\alias{EucDist}
\title{Compute Euclidean similarity}
\usage{
EucDist(data)
}
\arguments{
\item{data}{(M x N) dimensional inputmatrix. N is the number of objects M is the number of features describing each object.}
}
\value{
(N x N) dimensional normalized Euclidean similarity matrix.
}
\description{
Compute Euclidean similarity using a Gaussian kernel.
First, it normalize the input
then it transforms the distance to similarity using a Gaussian kernel
so that similar objects have larger values
}
\examples{
EucDist(matrix)

}
\references{
Hyundoo Jeong and Zhandong Liu

 PRIME: a probabilistic imputation method to reduce dropout effects in single cell RNA sequencing
}
\author{
Hyundoo Jeong
}
