% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{extract_rowname_components}
\alias{extract_rowname_components}
\title{Extract chr, start, end, and strand from row names
like "chr:start-end;strand"}
\usage{
extract_rowname_components(row_names)
}
\description{
Extract chr, start, end, and strand from row names
like "chr:start-end;strand"
}
