% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{extract_sample_label}
\alias{extract_sample_label}
\title{Extract sample label from input_basename (internal)}
\usage{
extract_sample_label(input_basename)
}
\description{
Returns the substring between 'pred_all_' and '_combined' if both exist;
otherwise returns 'PRIMEloci'.
}
