% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.r
\name{fillGaps}
\alias{fillGaps}
\title{Fill in missing scores across samples with zeros. Internal function used by \code{readRange}.}
\usage{
fillGaps(data)
}
\arguments{
\item{data}{A tibble of scores generated by \code{readRange}.}
}
\value{
A tibble with missing scores filled in with zeros.
}
\description{
Fill in missing scores across samples with zeros. Internal function used by \code{readRange}.
}
