% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIME.R
\name{FindNBR}
\alias{FindNBR}
\title{Find neighboring nodes in the network}
\usage{
FindNBR(dist_mat, qth = 0.9, th = 0.85)
}
\arguments{
\item{dist_mat}{Distance matrix}

\item{qth}{Threshold for the quantile of the similarity}

\item{th}{Hard threshold to select the neighborind nodes}
}
\value{
Adjacency matrix for the correspondence network
}
\description{
Find neighboring nodes in the network
}
\examples{
FindNBR(matrix, qth = 0.9, th = 0.85)

}
\references{
Hyundoo Jeong and Zhandong Liu

 PRIME: a probabilistic imputation method to reduce dropout effects in single cell RNA sequencing
}
\author{
Hyundoo Jeong
}
