% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{fingerPrint}
\alias{fingerPrint}
\title{Calculate the complexity of a CTSS dataset and generate data for a fingerprint plot.}
\usage{
fingerPrint(object, replicates, inputAssay = "counts", n = 1e+05)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} resulting from
\code{CAGEfightR::quantifyCTSSs()}.}

\item{replicates}{A character vector of replicate names to include in the analysis.}

\item{inputAssay}{A character string of the assay to use for the counts. Default is "counts".}

\item{n}{The number of points to include in the fingerprint plot. Default is 100,000.}
}
\value{
A data frame containing the sample names, raw counts, cumulative sums,
relative cumulative sums, and ranks for each sample.
}
\description{
This function calculates the complexity of a given CTSS dataset and generates a data
frame suitable for creating a fingerprint plot.
}
