% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getISO3}
\alias{getISO3}
\title{Retrieve ISO3-code of country}
\usage{
getISO3(countryname, name = FALSE)
}
\arguments{
\item{countryname}{Character string (required): Full name of the country}

\item{name}{Logical (optional): If TRUE, returns full name and alternative
names of returned country (may be useful to double-check that it is
the correct country)}
}
\value{
Character string with ISO3 code. Will also return full name
        if name=TRUE.
}
\description{
Retrieve ISO3-code of country
}
\examples{
getISO3("Afghanistan")
getISO3("Congo",name=TRUE)
}
