% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.r
\name{heatmapData}
\alias{heatmapData}
\title{Extract heatmap-like CTSS data for regions of fixed width
with optional sparse matrix output.}
\usage{
heatmapData(
  regions,
  data,
  column = "score",
  transform_fn = identity,
  sparse = FALSE,
  ...
)
}
\arguments{
\item{regions}{\code{GRanges} with regions to extract data for
(must be of the same size).}

\item{data}{\code{GRanges} with value to consider
in specified column (score).}

\item{column}{column in \code{mcols(data)} to consider.}

\item{transform_fn}{function applied on data for each region 
(e.g. to reduce dimension, defaults to identity).}

\item{sparse}{Logical indicating whether the output should be
in sparse matrix format. Default is FALSE.}

\item{...}{additional arguments to pass to \code{transform_fn}.}
}
\value{
list of matrices with data for each strand.
}
\description{
Extract heatmap-like CTSS data for regions of fixed width
with optional sparse matrix output.
}
