% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_noSparse.R
\name{heatmapData_noSparse}
\alias{heatmapData_noSparse}
\title{Extract heatmap-like CTSS data for regions of fixed width.}
\usage{
heatmapData_noSparse(
  regions,
  data,
  column = "score",
  transform_fn = identity,
  ...
)
}
\arguments{
\item{regions}{\code{GRanges} with regions to extract data for 
(must be of the same size).}

\item{data}{\code{GRanges} with value to consider in specified column 
(score).}

\item{column}{column in \code{mcols(data)} to consider.}

\item{transform_fn}{function applied on data for each region 
(e.g. to reduce dimension, defaults to identity).}

\item{...}{additional arguments to pass to \code{transform_fn}.}
}
\value{
list of matrices with data for each strand.
}
\description{
Extract heatmap-like CTSS data for regions of fixed width.
}
