% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqContent.R
\name{initiatorScore}
\alias{initiatorScore}
\title{Initiator classification of \code{GRanges} into YR, YC and other.}
\usage{
initiatorScore(object, bsg)
}
\arguments{
\item{object}{A \code{GRanges} or \code{StitchedGPos} object.}

\item{bsg}{The \code{BSgenome} reference to use.}
}
\value{
A \code{GRanges} object initiator classification.
}
\description{
Initiator classification of \code{GRanges} into YR, YC and other.
}
