% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{is_parallel_plan_working}
\alias{is_parallel_plan_working}
\title{Check if current future plan runs in parallel}
\usage{
is_parallel_plan_working()
}
\description{
Internal helper to test
whether the future plan executes with multiple PIDs.
}
