% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lifeTable}
\alias{lifeTable}
\title{Construct lifetable based on qx-column}
\usage{
lifeTable(qx = NULL, mx = NULL, agecohort = 0)
}
\arguments{
\item{qx}{Numeric vector (required): Age-specific probabilities of dying}

\item{agecohort}{Number (optional): Age at which cohort is started}
}
\value{
Data.table with lifetable
}
\description{
qx = age-specific probability of dying
}
\examples{

qx <- unlist(data.mortall[iso3=="AFG", as.character(0:100), with=F], use.names=F)
lifeTable(qx, 9)
}
