% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{local_maxima_decompose}
\alias{local_maxima_decompose}
\title{Decompose tag cluster into subclusters according to CTSS expression 
fraction of local maxima CTSS expression. Performs local summit 
decomposition for each local maxima separately in decreasing order of 
expression level. For each local summit decomposition, subclusters will be 
merged if within maxGap distance. If smoothPad>0, neighbouring non-zero 
CTSSs within smoothPad distance of CTSSs fulfilling the summit fraction 
criterion will also be included. Final subclusters within mergeDist bp 
will be merged.}
\usage{
local_maxima_decompose(
  views,
  fraction = 0.1,
  maximaDist = 20,
  maxGap = maximaDist,
  mergeDist = -1,
  smoothPad = 0
)
}
\arguments{
\item{views}{A \code{Views} object of tag clusters.}

\item{fraction}{The fraction of the summit CTSS expression that a CTSS
must have to be considered part of the tag cluster.}

\item{maximaDist}{The distance within which local maxima are considered.}

\item{maxGap}{The distance within which subclusters will be merged.}

\item{mergeDist}{The distance within which final decomposed subclusters 
will be merged.}

\item{smoothPad}{The distance within which neighbouring non-zero CTSSs will
be included.}
}
\value{
An \code{IRanges} object of decomposed tag clusters.
}
\description{
Decompose tag cluster into subclusters according to CTSS expression 
fraction of local maxima CTSS expression. Performs local summit 
decomposition for each local maxima separately in decreasing order of 
expression level. For each local summit decomposition, subclusters will be 
merged if within maxGap distance. If smoothPad>0, neighbouring non-zero 
CTSSs within smoothPad distance of CTSSs fulfilling the summit fraction 
criterion will also be included. Final subclusters within mergeDist bp 
will be merged.
}
