% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{modify_profile_rownames}
\alias{modify_profile_rownames}
\title{Modify row names to no-strand if both strands match}
\usage{
modify_profile_rownames(profiles, count_profiles)
}
\description{
Modify row names to no-strand if both strands match
}
