% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{monetary_to_number}
\alias{monetary_to_number}
\title{Convert monetary character-strings to numeric values}
\usage{
monetary_to_number(x)
}
\arguments{
\item{x}{Character string to convert}
}
\value{
Returns number with value, stripped from any currency symbols and
  thousand-seperators (i.e. "B#2,010.50" becomes 2010.5)
}
\description{
Convert monetary character-strings to numeric values
}
\examples{

monetary_to_number ("$2,200.20")

# Note that values using German or Dutch notation (i.e. using a comma to
  separate decimals and a dot to seperate thousands) are converted as well.
  monetary_to_number ("$2.200,20")

}
