% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{nonzero}
\alias{nonzero}
\title{nonzero function from DAPAR package (https://github.com/edyp-lab/DAPAR)
temporarily including a local version to avoid installation issues}
\usage{
nonzero(x)
}
\arguments{
\item{x}{A sparse matrix of class "dgCMatrix".}
}
\value{
A two-column matrix containing the indices of the non-zero elements in the input matrix.
}
\description{
nonzero function from DAPAR package (https://github.com/edyp-lab/DAPAR)
temporarily including a local version to avoid installation issues
}
