% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeBySizeFactors}
\alias{normalizeBySizeFactors}
\title{Normalize expression counts by size factors.}
\usage{
normalizeBySizeFactors(
  object,
  sizeFactors,
  inputAssay = "counts",
  outputAssay = "normalized"
)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object.}

\item{sizeFactors}{The size factors to use for the normalization.}

\item{inputAssay}{The assay to use.}

\item{outputAssay}{The assay to use for the output.}
}
\value{
A \code{SummarizedExperiment} object with the normalized counts.
}
\description{
Normalize expression counts by size factors.
}
