% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_functions.r
\name{OutputGavi}
\alias{OutputGavi}
\title{OutputGavi}
\usage{
OutputGavi(DT, age_stratified = TRUE, calendar_year = FALSE,
  gavi_template = -1)
}
\arguments{
\item{gavi_template}{}
}
\value{
#
}
\description{
OutputGavi
}
\examples{
#
}
