% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{OutputVimc}
\alias{OutputVimc}
\title{Formatting output for VIMC Montagu}
\usage{
OutputVimc(DT, age_stratified = TRUE, calendar_year = FALSE,
  vimc_template = -1)
}
\arguments{
\item{DT}{data table with results}

\item{age_stratified}{logical, whether output should be stratified by age}

\item{calendar_year}{logical, whether output should be given by calendar year
of event OR by year of birth of cohort}

\item{vimc_template}{data table with template file downloaded from montagu}
}
\value{
#
}
\description{
\code{OutputVimc} takes result of BatchRun and outputs it in format to be
  uploaded to VIMC Montagu.
}
\examples{
#
}
