% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_batch_heatmapData}
\alias{plc_batch_heatmapData}
\title{Batch wrapper for heatmapData that preserves input order}
\usage{
plc_batch_heatmapData(regions, data, batch_size = 1000, sparse = TRUE, ...)
}
\arguments{
\item{regions}{A GRanges object of equal-width regions}

\item{data}{A GRanges object with CTSS signal}

\item{batch_size}{Number of regions to process per batch}

\item{sparse}{Logical, whether to return sparse matrix (default = TRUE)}

\item{...}{Additional arguments passed to `heatmapData`}
}
\value{
A list of matrices by strand with rows in the same order as `regions`
}
\description{
Batch wrapper for heatmapData that preserves input order
}
