% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_check_multisession_parallel}
\alias{plc_check_multisession_parallel}
\title{Check if multisession is truly parallel}
\usage{
plc_check_multisession_parallel(
  num_workers = 2,
  timeout_sec = 15,
  sleep_sec = 2,
  overlap_threshold = 6,
  runtime_threshold = 10
)
}
\description{
Check if multisession is truly parallel
}
