% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_check_python_dependencies}
\alias{plc_check_python_dependencies}
\title{Check Required Python Dependencies}
\usage{
plc_check_python_dependencies(required_packages)
}
\arguments{
\item{required_packages}{Named list of required packages
with minimum versions.}
}
\description{
Checks that all required Python packages are installed and meet
the specified minimum version requirements. Raises an error
if any package is missing or incompatible.
}
