% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_create_granges_from_bed}
\alias{plc_create_granges_from_bed}
\title{Create a GRanges object from a BED data.table}
\usage{
plc_create_granges_from_bed(bed_file)
}
\description{
This function converts a `data.table`
containing BED file data into a `GRanges` object.
If `thickStart` and `thickEnd` are present,
they are converted to 1-based `IRanges` and added as metadata.
}
