% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_create_output_dir}
\alias{plc_create_output_dir}
\title{Create the output directory if it doesn't exist}
\usage{
plc_create_output_dir(output_dir)
}
\description{
Create the output directory if it doesn't exist
}
