% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_detect_parallel_plan}
\alias{plc_detect_parallel_plan}
\title{Detect optimal parallel backend (multisession or callr)}
\usage{
plc_detect_parallel_plan(
  num_workers = 2,
  startup_tolerance_sec = 90,
  sleep_sec = 2
)
}
\description{
Detect optimal parallel backend (multisession or callr)
}
