% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_facet_prediction_to_rse}
\alias{plc_facet_prediction_to_rse}
\title{Convert a directory of facet predictions to a RangedSummarizedExperiment}
\usage{
plc_facet_prediction_to_rse(
  facet_prediction_dir,
  postprocess_partial_name = "pred_all"
)
}
\description{
Convert a directory of facet predictions to a RangedSummarizedExperiment
}
