% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_focal_prediction_to_rse}
\alias{plc_focal_prediction_to_rse}
\title{Convert a directory of focal predictions to a RangedSummarizedExperiment}
\usage{
plc_focal_prediction_to_rse(
  focal_prediction_dir,
  postprocess_partial_name = "pred_all"
)
}
\description{
Convert a directory of focal predictions to a RangedSummarizedExperiment
}
