% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_get_ctss_from_bw}
\alias{plc_get_ctss_from_bw}
\title{Extract CAGE Transcription Start Sites (CTSSs) from BigWig Files}
\usage{
plc_get_ctss_from_bw(
  dir_cage_bw,
  design_matrix,
  genome_info = NULL,
  drop_chr = NULL,
  keep_chr = NULL
)
}
\arguments{
\item{dir_cage_bw}{A character string specifying the directory containing
the BigWig files for both plus and minus strands.}

\item{design_matrix}{A data frame that must contain columns named
\code{"BigWigPlus"} and \code{"BigWigMinus"}, which specify the paths
to the respective BigWig files.}

\item{genome_info}{An optional \code{Seqinfo} object containing genome
information. If provided, it allows for the use of `drop_chr`
and `keep_chr` parameters.}

\item{drop_chr}{An optional character vector of chromosome names to be
excluded from the analysis. Requires \code{genome_info} to be specified.}

\item{keep_chr}{An optional character vector of chromosome names to be
included in the analysis. Requires \code{genome_info} to be specified.}
}
\value{
A \code{CTSS} object containing the quantified and normalized
  counts of transcription start sites, including pooled counts.
}
\description{
This function extracts CTSSs from CAGE data stored in BigWig files,
as specified in the provided design matrix. It allows for optional
genomic filtering based on a provided `Seqinfo` object and can exclude
or include specific chromosomes.
}
\examples{
\dontrun{
ctss_result <- plc_get_ctss_from_bw("path/to/bigwig/files",
                                    design_matrix,
                                    genome_info,
                                    drop_chr = c("chrM"))
}
}
