% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_load_bed_file}
\alias{plc_load_bed_file}
\title{Load a BED file and validate its columns}
\usage{
plc_load_bed_file(input_bed)
}
\arguments{
\item{input_bed}{Character. The path to the input BED file.}
}
\value{
A `data.table` containing the BED file data.
}
\description{
This function reads a BED file into a `data.table`
and checks that it contains the required columns:
'chrom', 'chromStart', 'chromEnd', 'strand', and 'score'.
}
