% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_postprocessed_bed_to_rds}
\alias{plc_postprocessed_bed_to_rds}
\title{Convert a BED file to a GRanges object and save as RDS}
\usage{
plc_postprocessed_bed_to_rds(
  bed_file,
  colnames = c("chrom", "chromStart", "chromEnd", "name", "score", "strand",
    "thickStart", "thickEnd")
)
}
\arguments{
\item{bed_file}{A character string specifying the path to the input BED file.}

\item{colnames}{A character vector of column names to assign to the BED file.
Defaults to standard BED fields: `chrom`, `chromStart`, `chromEnd`,
`name`, `score`, `strand`, `thickStart`, `thickEnd`.}
}
\value{
No return value. A `.rds` file
containing the GRanges object is saved to disk.
}
\description{
This function reads a BED file, assigns column names,
converts the data to a `GRanges` object
using `create_granges_from_bed()`,
and saves the result as an `.rds` file.
}
