% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_set_parallel_plan}
\alias{plc_set_parallel_plan}
\title{Set a robust parallel plan using multisession or callr
Use multisession if workers >= 2.
If multisession is not truly parallel, fall back to callr.
Always warn if callr is used.
Do not fall back to sequential if callr fails.}
\usage{
plc_set_parallel_plan(num_workers = 1)
}
\description{
Set a robust parallel plan using multisession or callr
Use multisession if workers >= 2.
If multisession is not truly parallel, fall back to callr.
Always warn if callr is used.
Do not fall back to sequential if callr fails.
}
