% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_setup_log_target}
\alias{plc_setup_log_target}
\title{Set up logging to console or file}
\usage{
plc_setup_log_target(log, outdir)
}
\arguments{
\item{log}{Logical or NULL. If NULL, log to console; otherwise,
log to file.}

\item{outdir}{Character. Path to the output directory
where logs will be stored.}
}
\value{
A connection or file path to be used as the logging target.
}
\description{
Set up logging to console or file
}
