% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_trylog}
\alias{plc_trylog}
\title{Capture and log messages, warnings, and errors from a code block}
\usage{
plc_trylog(expr, log_file, print_console = TRUE)
}
\arguments{
\item{expr}{Code to evaluate.}

\item{log_file}{Path to the log file.}

\item{print_console}{Logical. Whether to also print output to console.}
}
\value{
The result of the evaluated expression, or NULL on error.
}
\description{
Capture and log messages, warnings, and errors from a code block
}
