% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_validate_tc_object}
\alias{plc_validate_tc_object}
\title{Validate a TC object}
\usage{
plc_validate_tc_object(tc_object, ctss_rse, ext_dis = 200)
}
\arguments{
\item{tc_object}{A `GenomicRanges::GRanges` or
`GenomicRanges::GRangesList` object.}

\item{ctss_rse}{The original CTSS dataset (for length and naming validation).}

\item{ext_dis}{An integer value for extension distance. Default is 200.}
}
\value{
TRUE if validation passes; otherwise, throws an error.

#' @examples
# Validate an existing TC object:
plc_validate_tc_object(tc_object, ctss_rse)
}
\description{
This function ensures that a given `GenomicRanges::GRanges` or
`GenomicRanges::GRangesList` object meets the expected criteria.
}
