% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.r
\name{plotTracks}
\alias{plotTracks}
\title{Plot tracks for a specified genomic range using ggplot2.}
\usage{
plotTracks(data, range)
}
\arguments{
\item{data}{A tibble of scores for the specified genomic range, generated by \code{readRange}.}

\item{range}{A \code{GRanges} object specifying the genomic range to plot}
}
\value{
A ggplot object representing the tracks for the specified genomic range.
}
\description{
Plot tracks for a specified genomic range using ggplot2.
}
