% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{poolReplicates}
\alias{poolReplicates}
\title{Pools expression signal across replicates}
\usage{
poolReplicates(object, replicates, inputAssay = "counts")
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object.}

\item{replicates}{A \code{factor} indicating the replicate ID of each sample.}

\item{inputAssay}{The assay to use}
}
\value{
A \code{SummarizedExperiment} expression pooled across replicates and merged column information
}
\description{
Pools expression signal across replicates
}
