% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{prep_profile_dir}
\alias{prep_profile_dir}
\title{Prepare Directory Structure for Profile Output}
\usage{
prep_profile_dir(
  output_dir = ".",
  profile_dir_name = "profile_output",
  dir_name = c("metadata", "profiles", "profiles_subtnorm", "predictions")
)
}
\description{
Prepare Directory Structure for Profile Output
}
