% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{prime}
\alias{prime-package}
\alias{prime}
\title{prime: Papillomavirus Rapid Interface for Modelling and Economics (PRIME).}
\description{
PRIME stands for “Papillomavirus Rapid Interface for Modelling and Economics”.
The R package is based of the spreadsheet-based tool (see http://primetool.org).
}
\section{PRIME provides estimates of}{

\describe{
\item{The magnitude of the burden of cervical cancer.}{}
\item{The impact of introducing HPV vaccination for girls prior to sexual debut.}{}
\item{Healthcare costs incurred as a result of cervical cancer treatment.}{}
\item{Costs associated with vaccination.}{}
\item{Long-term savings which may result from a vaccination program.}{}
}
}

