% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci.R
\name{PRIMEloci}
\alias{PRIMEloci}
\title{PRIMEloci function}
\usage{
PRIMEloci(
  ctss_rse,
  tc_object = NULL,
  outdir,
  ext_dis = 200,
  save_tc = TRUE,
  tc_object_name = "tc_grl.rds",
  sld_window = TRUE,
  sld_by = 20,
  sld_object_name = "sld_tc_grl.rds",
  keep_tmp = TRUE,
  num_cores = NULL
)
}
\description{
PRIMEloci function
}
