% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_focal.R
\name{PRIMEloci_focal}
\alias{PRIMEloci_focal}
\title{Run the PRIMEloci focal Pipeline on CTSS and Identified Regions}
\usage{
PRIMEloci_focal(
  ctss_rse,
  tc_gr,
  python_path = NULL,
  num_cores = NULL,
  keep_tmp = FALSE,
  log_dir = NULL
)
}
\arguments{
\item{ctss_rse}{A `RangedSummarizedExperiment` object representing CTSS data.}

\item{tc_gr}{A `GRanges` object representing identified regions.
Regions will be extended to 401 bp width if needed.}

\item{python_path}{Character path to the Python binary
in the desired environment. Default is NULL.}

\item{num_cores}{Optional integer specifying the number of CPU cores
to use for parallel steps.}

\item{keep_tmp}{Logical. If `TRUE`, temporary files and folders
will be retained.
  Default is `FALSE`.}

\item{log_dir}{Optional path to save a log file.
If `NULL`, logs are printed to the console.}
}
\value{
A `GRanges` object if one sample was processed,
  or a `GRangesList` object if multiple samples were processed.
}
\description{
This function executes a partial PRIMEloci pipeline,
which includes the steps of validating and extending existing regions—
such as those predicted from pooled PRIMEloci data—
followed by profile generation, model prediction, and BED result import.
}
\details{
This function supports downstream analysis by:
- Validating and, if necessary, extending input regions to 401 bp
- Generating normalized CTSS-based profiles per sample
- Running predictions using a pre-trained PRIMEloci model via Python
- Importing the prediction results from BED files

Temporary files are stored in a subdirectory of `tempdir()`
and removed unless `keep_tmp = TRUE`.
}
