% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_functions.r
\name{propSexDebut}
\alias{propSexDebut}
\title{propSexDebut}
\usage{
propSexDebut(age, country_iso3)
}
\arguments{
\item{country_iso3}{}
}
\value{
#returns proportion of girls in country that has sexually debuted
}
\description{
propSexDebut
}
\examples{
#.
}
