% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergent.R
\name{quantifyClustersOlap}
\alias{quantifyClustersOlap}
\title{Quantify the expression of disjoint tag clusters.}
\usage{
quantifyClustersOlap(object, clusters, inputAssay)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} resulting from
\code{CAGEfightR::quantifyCTSSs()}.}

\item{clusters}{A \code{GRanges} object containing the tag clusters.}

\item{inputAssay}{The assay to use for quantification.}
}
\value{
A \code{SummarizedExperiment} object containing the quantified
tag clusters.
}
\description{
Quantify the expression of disjoint tag clusters.
}
