% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergent.R
\name{quantifyDivergentLoci}
\alias{quantifyDivergentLoci}
\title{Quantify the expression of divergent loci (identified by 
\code{divergentLoci}).}
\usage{
quantifyDivergentLoci(
  loci,
  ctss,
  inputAssay = "counts",
  requireDisjoint = TRUE
)
}
\arguments{
\item{loci}{A \code{GRanges} object containing the divergent loci.}

\item{ctss}{A \code{RangedSummarizedExperiment} resulting from
\code{CAGEfightR::quantifyCTSSs()}.}

\item{inputAssay}{The assay to use for quantification.}

\item{requireDisjoint}{Logical indicating whether to require divergent 
loci to be disjoint.}
}
\value{
A \code{SummarizedExperiment} object containing the quantified
divergent loci.
}
\description{
Quantify the expression of divergent loci (identified by 
\code{divergentLoci}).
}
