% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergent.R
\name{quantifyStrandwiseDivergentLoci}
\alias{quantifyStrandwiseDivergentLoci}
\title{import BiocGenerics}
\usage{
quantifyStrandwiseDivergentLoci(
  loci,
  ctss,
  inputAssay = "counts",
  requireDisjoint = TRUE
)
}
\arguments{
\item{loci}{A \code{GRanges} object containing the divergent loci.}

\item{ctss}{A \code{RangedSummarizedExperiment} resulting from
\code{CAGEfightR::quantifyCTSSs()}.}

\item{inputAssay}{The assay to use for quantification.}

\item{requireDisjoint}{Logical indicating whether to require divergent
loci to be disjoint.}
}
\description{
import BiocGenerics
}
