% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.r
\name{readRange}
\alias{readRange}
\title{Read bigWig files for a specified genomic range and return a tibble of scores.}
\usage{
readRange(files, range)
}
\arguments{
\item{files}{A character vector of bigWig file paths to read.}

\item{range}{A \code{GRanges} object specifying the genomic range to read from the bigWig files.}
}
\value{
A tibble of scores for the specified genomic range.
}
\description{
Read bigWig files for a specified genomic range and return a tibble of scores.
}
