% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_functions.r
\name{RegisterBatchData}
\alias{RegisterBatchData}
\title{RegisterBatchData}
\usage{
RegisterBatchData(coverage_data, reporting_years = -1, force = FALSE)
}
\arguments{
\item{coverage_data}{- DT with columns country_code, year (of vaccination), age_first, age_last, coverage}

\item{reporting_years}{- numeric_vector for which years should be reported}

\item{force}{}
}
\value{
#
}
\description{
RegisterBatchData
}
\examples{
#
}
