% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RegisterBatchDataVimc}
\alias{RegisterBatchDataVimc}
\title{Creates .data.batch for running multiple birth cohorts (VIMC runs)}
\usage{
RegisterBatchDataVimc(vimc_coverage, vimc_template, use_campaigns,
  use_routine, restrict_to_coverage_data = FALSE, force = FALSE,
  psa = 0)
}
\arguments{
\item{vimc_coverage}{data table with coverage estimates as downloaded from
VIMC montagu}

\item{vimc_template}{data table with reporting template as downloaded from
VIMC montagu}

\item{use_campaigns}{logical, whether campaigns as stated in coverage files
should be modelled}

\item{use_routine}{logical, whether routine vaccination as stated in coverage
file should be modelled}

\item{restrict_to_coverage_data}{logical, whether the first birth-cohort
should be the first cohort that is mentioned in the coverage data.
  If TRUE, restrict to coverage data.
  If FALSE, restrict to cohorts provided in vimc_template.}

\item{force}{logical, whether .data.batch should be overwritten if it already
exists}

\item{psa}{integer, indicating how many runs for probabilistic sensitivity
analysis (PSA). 0 to run no PSA.}
}
\value{
batch data of cohorts with vaccination coverage
}
\description{
Creates .data.batch which is used when running/looping over multiple birth
  cohorts ( runCohort() ) at once. Similar to RegisterBatchData, but for when
  we make runs for VIMC.
}
\details{
.data.batch is based on the data.table (DT) coverage_data, which is a DT with
  columns country_code (ISO3), year (of vaccination),
  age_first (age at vaccination), age_last (age at vaccination),
  coverage (in proportion, for all the age groups specified).
}
\examples{
#
}
