% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{rmSingletons}
\alias{rmSingletons}
\title{Remove singleton counts from a \code{SummarizedExperiment} object. 
Singleton counts are those that have a value of 1 in the specified assay. 
This function replaces those values with 0 and stores the modified assay in a new column.}
\usage{
rmSingletons(rse, inputAssay = "counts", outputAssay = "counts.noSingletons")
}
\arguments{
\item{rse}{A \code{SummarizedExperiment} object.}

\item{inputAssay}{The name of the assay to use.}

\item{outputAssay}{The name of the assay to store the modified counts. Default is "counts.noSingletons".}
}
\value{
A \code{SummarizedExperiment} object with the modified assay.
}
\description{
Remove singleton counts from a \code{SummarizedExperiment} object. 
Singleton counts are those that have a value of 1 in the specified assay. 
This function replaces those values with 0 and stores the modified assay in a new column.
}
