% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_PRIMEloci_facet_example.R
\name{run_PRIMEloci_facet_example}
\alias{run_PRIMEloci_facet_example}
\title{Run an Example of the PRIMEloci Facet Pipeline}
\usage{
run_PRIMEloci_facet_example(
  python_path = NULL,
  log_dir = NULL,
  keep_tmp = FALSE,
  ...
)
}
\arguments{
\item{python_path}{Character path to the Python binary
in the desired environment. Default is NULL.}

\item{log_dir}{Optional character path to save the log file.
Default is `"/Users/natsudanav/Desktop/PRIMEloci_facet.log"`.}

\item{keep_tmp}{Logical. If `TRUE`, temporary files and folders are retained.
Default is `FALSE`.}
}
\value{
A `GRanges` object or a `GRangesList` object,
  depending on the number of samples in the example dataset.
}
\description{
This function demonstrates how to run the `PRIMEloci_facet()` pipeline
using example CTSS and region data bundled with the `PRIME` package.
It loads pre-packaged test data, runs the pipeline, and returns the result.
}
