% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RunCohort}
\alias{RunCohort}
\title{Run PRIME for a single birth-cohort}
\usage{
RunCohort(lifetab, cohort, incidence, mortality_cecx, prevalence, agevac,
  coverage, campaigns, vaccine_efficacy_nosexdebut,
  vaccine_efficacy_sexdebut, daly.canc.diag, daly.canc.seq,
  daly.canc.terminal, cost_cancer, disc.cost = 0.03, disc.ben = 0.03,
  discounting = TRUE, country_iso3 = NULL, run_country = FALSE)
}
\arguments{
\item{lifetab}{Data.table: The life-table for this cohort. Can be created using the lifeTable() function.}

\item{cohort}{Number: The cohort-size of this birth-cohort at the time where the lifetable starts.}

\item{incidence}{Numeric vector: Age-specific CeCx(16/18) incidence-rates.}

\item{mortality_cecx}{Numeric vector: Age-specific CeCx(16/18) mortality-rates.}

\item{prevalence}{Numeric vector: Age-specific CeCx(16/18) prevalence rates (5-year prevalence) -- referring to people who are alive within 5 years of diagnosis.}

\item{agevac}{Number: Age at which the cohort is vaccinated.}

\item{coverage}{Number: Proportion of the cohort that will receive a vaccination.}

\item{campaigns}{List or number: MAC cohort-vaccinations (needs to be changed).}

\item{daly.canc.diag}{Number: daly-weight for cancer diagnosis.}

\item{daly.canc.seq}{Number: daly-weight for cancer...}

\item{daly.canc.terminal}{Number: daly-weight for death from cancer.}

\item{cost_cancer}{Number: total per capita cost of cancer.}

\item{disc.cost}{Number (optional): discounting for cancer cost.}

\item{disc.ben}{Number (optional): discounting for...}

\item{discounting}{Logical: should discounting be applied?}

\item{vaccine_efficacy}{Number: proportion indicating vaccine-efficacy.}
}
\value{
Returns a data.table with size of the birth-cohort and age-specific incidence-rates,
    mortality-rates, years-of-life-lost, years-of-healthy-life-lost, and cancer-costs before and after vaccination
}
\description{
Runs PRIME for one birth-cohort. Usually called by another function such as RunCountry().
}
\examples{
lifetab <- lifeTable(unlist(data.mortall[iso3=="AFG", as.character(0:100), with=F], use.names=F), 9)
cohort <- unlist(data.popproj[iso3=="AFG", "2020"], use.names=F)
incidence <- unlist(data.incidence[iso3=="AFG",as.character(0:100),with=F],use.names=F)
mortality_cecx <- unlist(data.mortall[iso3=="AFG",as.character(0:100),with=F],use.names=F)
agevac <- 9
coverage <- 0.8
campaigns <- -1
vaccine_efficacy <- 0.95
daly.canc.diag <- 0.0003
daly.canc.seq <- 0.002
daly.canc.terminal <- 0.1
cost.cancer <- 100
RunCohort(lifetab, cohort, incidence, mortality_cecx, prevalence, agevac, coverage, campaigns, vaccine_efficacy, daly.canc.diag, daly.canc.seq, daly.canc.terminal, daly.canc.terminal, cost.cancer)

}
