% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RunCountry}
\alias{RunCountry}
\title{Run PRIME for a specific country}
\usage{
RunCountry(country_iso3, vaceff_beforesexdebut = 1,
  vaceff_aftersexdebut = 0, disc.cost = 0.03, disc.ben = 0.03,
  cov = 1, agevac = 10, agecohort = 10, cohort = -1,
  canc.cost = "unadj", canc.inc = "2018", daly.canc.diag = 0.288,
  daly.canc.control = 0.049, daly.canc.metastatic = 0.451,
  daly.canc.terminal = 0.54, sens = -1, unwpp_mortality = TRUE,
  year_born = -1, year_vac = -1, campaigns = -1,
  analyseCosts = FALSE, discounting = TRUE, run_batch = FALSE,
  psadat = -1, disability.weights = "gbd_2017")
}
\arguments{
\item{country_iso3}{Character string (required): ISO3 code of the country}

\item{disc.cost}{Number (optional): Discounting for costs (only if discounting=TRUE)}

\item{disc.ben}{Number (optional): Discounting for ... (only if discounting=TRUE)}

\item{cov}{Number (optional): Proportion with routine coverage}

\item{agevac}{Integer (optional): Target age for HPV vaccination}

\item{agecohort}{Integer (optional): Reference age for cohort-size (only used when 'cohort' is not provided)}

\item{cohort}{Integer (optional): Cohort-size. -1 if unknown}

\item{canc.cost}{Character (optional): Is cost of cancer adjusted ("adj") or not ("unadj")}

\item{canc.inc}{Integer (optional): Reference year for cancer incidence rates (Globocan: 2018 or 2012)}

\item{daly.canc.diag}{Number (optional): Daly weight for cancer diagnosis}

\item{daly.canc.control}{Number: daly-weight for controlled phase of cervical cancer}

\item{daly.canc.metastatic}{Number: daly-weight for metastatic phase of cervical cancer}

\item{daly.canc.terminal}{Number (optional): Daly weight for cancer death}

\item{sens}{Numeric-vector (optional): Specific values to be used in a PSA. -1 if PSA's are not used}

\item{unwpp_mortality}{Logical (optional): If TRUE, uses year-specific UNWPP mortality estimates to construct life-tables. If FALSE, use WHO based mortality estimates}

\item{year_born}{Integer (optional): Year in which cohort is born}

\item{year_vac}{Integer (optional): Year in which cohort is vaccinated}

\item{campaigns}{List (optional): Multi-Age-Cohort campaigns (needs to be changed)}

\item{analyseCosts}{Logical (optional): If FALSE, returns result from RunCohort() function. If TRUE, runs analyseCosts() with country-specific results}

\item{discounting}{Logical (optional): If TRUE, runs analysis undiscounted and discounted. If FALSE, only uses undiscounted}

\item{disability.weights}{character, disability weights for cervical cancer from GBD 2017 or GBD 2001}

\item{vaceff}{Number (optional): Proportion indicating vaccine-efficacy}
}
\value{
data.table with country-specific results of HPV vaccination. Returns cost-analysis if analyseCosts=TRUE
}
\description{
Runs RunCohort() using country-specific estimates.
If year_born and year_vac are not provided, assumes vaccination occurs in the current year.
}
\examples{
RunCountry("AFG")
RunCountry("AFG", year_vac=2020, agevac=10, cov=0.75, vaceff=0.88)
RunCountry("AFG", year_vac=2020, agevac=10, cov=0.75, vaceff=0.88, analyseCosts=TRUE)
}
