% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{set_parallel_plan}
\alias{set_parallel_plan}
\title{Set a robust parallel plan using multisession or callr}
\usage{
set_parallel_plan(method_to_use, num_workers)
}
\description{
Set a robust parallel plan using multisession or callr
}
