% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{setup_tmp_dir}
\alias{setup_tmp_dir}
\title{Set up temporary directory under output directory}
\usage{
setup_tmp_dir(output_dir)
}
\description{
Set up temporary directory under output directory
}
