% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsampleProportion}
\alias{subsampleProportion}
\title{Subsample a \code{SummarizedExperiment} to a proportion of its sequencing 
depth. Useful for saturation analyses.}
\usage{
subsampleProportion(object, inputAssay = "counts", proportion)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object.}

\item{inputAssay}{The assay to use.}

\item{proportion}{The proportion of the sequencing depth to subsample.}
}
\value{
A \code{SummarizedExperiment} object with the subsampled counts.
}
\description{
Subsample a \code{SummarizedExperiment} to a proportion of its sequencing 
depth. Useful for saturation analyses.
}
