% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsampleTarget}
\alias{subsampleTarget}
\title{Subsample a \code{SummarizedExperiment} to a target sequencing depth.}
\usage{
subsampleTarget(object, inputAssay = "counts", target)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object.}

\item{inputAssay}{The assay to use.}

\item{target}{The target sequencing depth.}
}
\value{
A \code{SummarizedExperiment} object with the subsampled counts.
}
\description{
Subsample a \code{SummarizedExperiment} to a target sequencing depth.
}
