% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{summit_decompose}
\alias{summit_decompose}
\title{Decompose tag cluster into subclusters according to CTSS expression
fraction of summit CTSS expression. Subclusters within mergeDist bp
will be merged.}
\usage{
summit_decompose(views, fraction = 0.1, mergeDist = 20)
}
\arguments{
\item{views}{A \code{Views} object of tag clusters.}

\item{fraction}{The fraction of the summit CTSS expression that a CTSS
must have to be considered part of the tag cluster.}

\item{mergeDist}{The distance within which subclusters will be merged.}
}
\value{
An \code{IRanges} object of decomposed tag clusters.
}
\description{
Decompose tag cluster into subclusters according to CTSS expression
fraction of summit CTSS expression. Subclusters within mergeDist bp
will be merged.
}
