% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{TPMnormalizeBySizeFactors}
\alias{TPMnormalizeBySizeFactors}
\title{TPM normalize expression counts, considering size factors.}
\usage{
TPMnormalizeBySizeFactors(
  object,
  sizeFactors,
  inputAssay = "counts",
  outputAssay = "TPM"
)
}
\arguments{
\item{object}{A \code{SummarizedExperiment} object.}

\item{sizeFactors}{The size factors to use for the normalization.}

\item{inputAssay}{The assay to use.}

\item{outputAssay}{The assay to use for the output.}
}
\value{
A \code{SummarizedExperiment} object with the TPM normalized counts.
}
\description{
TPM normalize expression counts, considering size factors.
}
