% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_functions.r
\name{warningSexDebut}
\alias{warningSexDebut}
\title{Warning for missing sexual debut data}
\usage{
warningSexDebut(age = 999, country_iso3 = "ZZZ", remove_file = FALSE,
  return_file = FALSE)
}
\arguments{
\item{return_file}{}
}
\value{
processes warnings regarding sex debuts
}
\description{
Provide warning if sexual debut data for country is unavailable
}
\examples{
#.
}
