% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{write_granges_to_bed_coreovlwithd}
\alias{write_granges_to_bed_coreovlwithd}
\title{Write a GRanges object to a BED file for coreovlwith-d.}
\usage{
write_granges_to_bed_coreovlwithd(
  gr,
  output_dir,
  input_basename,
  output_basename
)
}
\description{
This function converts a GRanges object to a data frame
and writes it to a BED file, including metadata such as
thick position and maximum score.
}
