% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writelog}
\alias{writelog}
\title{Simulation log reporting}
\usage{
writelog(logname, x)
}
\arguments{
\item{logname}{log filename}

\item{x}{message of simulation run}
}
\value{
None
}
\description{
Appends message of simulation run (\code{x}) to log file (\code{logname}).
}
\examples{
#

}
