% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BatchRun}
\alias{BatchRun}
\title{Run multiple cohorts in a batch}
\usage{
BatchRun(countries = -1, coverage = -1, agevac = -1,
  agecohort = -1, canc.inc = "2018", daly.canc.diag = 0.288,
  daly.canc.control = 0.049, daly.canc.metastatic = 0.451,
  daly.canc.terminal = 0.54, sens = -1, unwpp_mortality = TRUE,
  year_born = -1, year_vac = -1, runs = 1,
  vaccine_efficacy_beforesexdebut = 1,
  vaccine_efficacy_aftersexdebut = 0, log = -1,
  by_calendaryear = FALSE, use_proportions = TRUE,
  analyseCosts = FALSE, canc.cost = "unadj", psa = 0,
  psa_vals = ".data.batch.psa", disability.weights = "gbd_2017",
  wb.indicator = "NY.GDP.PCAP.PP.CD", wb.year = 2017)
}
\arguments{
\item{countries}{ignore, read from .data.batch}

\item{coverage}{ignore, read from .data.batch}

\item{agevac}{ignore, read from .data.batch}

\item{agecohort}{ignore, read from .data.batch}

\item{canc.inc}{year from where incidence data is read (2018; old data: 2008/2012) -- with updated 2018 Globocan data, DALY weights from GBD, and DALY estimation based on prevalence instead of age of incidence, only 2018 is valid and 2008/2012 is not sensible}

\item{daly.canc.diag}{disability weight of diagnosis and primary therapy phase of cervical cancer}

\item{daly.canc.control}{Number: daly-weight for controlled phase of cervical cancer}

\item{daly.canc.metastatic}{Number: daly-weight for metastatic phase of cervical cancer}

\item{daly.canc.terminal}{disability weight of terminal phase of cervical cancer}

\item{sens}{ignore, does not do anything anymore}

\item{unwpp_mortality}{logical, whether to create lifetables based on UNWPP mx estimates or WHO data}

\item{year_born}{ignore}

\item{year_vac}{ignore}

\item{runs}{ignore}

\item{vaccine_efficacy_beforesexdebut}{vaccine efficacy before sexual debut}

\item{vaccine_efficacy_aftersexdebut}{vaccine efficacy after sexual debut}

\item{log}{name of log file}

\item{by_calendaryear}{logical, output values by calendar year or by year of birth cohort}

\item{use_proportions}{logical, output data as rates per capita or in totals}

\item{analyseCosts}{logical, directly run cost-effectiveness analysis on output or not}

\item{canc.cost}{Character (optional): Is cost of cancer adjusted
("adj" for International $) or not ("unadj" for US$)}

\item{psa}{integer, number of runs for probabilistic sensitivity analysis (PSA)}

\item{psa_vals}{data table with values to use in probabilistic sensitivity analysis, usually .data.batch.psa, generated by RegisterBatchData* functions (currently only RegisterBatchdataGavi)}

\item{disability.weights}{character, disability weights for cervical cancer from GBD 2017 or GBD 2001}

\item{wb.indicator}{character, World Bank indicator for GDP/GNI per capita in I$/US$ and current/constant data}

\item{wb.year}{numeric, year of the World Bank indicator value}
}
\value{
Returns combined results
}
\description{
Runs multiple cohorts in one batch, based on the data in .data.batch
}
\examples{
#
}
