% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{cumulativeFractionAroundLoci}
\alias{cumulativeFractionAroundLoci}
\title{Calculate the auto correlation of CTSSs over tag clusters according to 
pooled values and the cross correlation with opposite strand CTSSs}
\usage{
cumulativeFractionAroundLoci(loci, ctss, max_dist = 1000)
}
\arguments{
\item{loci}{A \code{GRanges} object of loci to focus on.}

\item{ctss}{A \code{RangedSummarizedExperiment} object containing the CTSSs.}

\item{max_dist}{The number of base pairs to consider around each locus.}
}
\value{
A data frame containing the cumulative fraction of CTSSs around
the loci, for each sample.
}
\description{
Calculate the auto correlation of CTSSs over tag clusters according to 
pooled values and the cross correlation with opposite strand CTSSs
}
