% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.costcecx}
\alias{data.costcecx}
\title{Cost of cervical cancer treatment}
\format{A data table with 194 observations of 4 variables.
\describe{
  \item{country}{Country name}
  \item{cancer_cost}{cost per cancer episode, in $US}
  \item{cancer_cost_adj}{cost per cancer episode -- adjusted, in international/PPP $}
  \item{iso3}{ISO3 country code}
}}
\usage{
data.costcecx
}
\description{
A dataset containing the cost of cervical cancer treatment.
}
\keyword{datasets}
