% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.hpv_distribution}
\alias{data.hpv_distribution}
\title{Relative contribution of HPV 16/18/31/33/45/52/58 in ICC HPV-positive cases}
\format{
A data table with 249 observations of 12 variables.
\describe{
  \item{Global}{World}
  \item{Region}{UN region}
  \item{Subregion}{UN subregion}
  \item{Intermediate_region}{UN intermediate region}
  \item{Country}{Country name}
  \item{iso3}{ISO3 country code}
  \item{hpv_4v}{Relative contribution (\%) of HPV 16/18 in ICC HPV-positive cases (mean)}
  \item{hpv_4v_low}{Relative contribution (\%) of HPV 16/18 in ICC HPV-positive cases (lower bound of 95\% uncertainty interval)}
  \item{hpv_4v_high}{Relative contribution (\%) of HPV 16/18 in ICC HPV-positive cases (upper bound of 95\% uncertainty interval)}
  \item{hpv_9v}{Relative contribution (\%) of HPV 16/18/31/33/45/52/58 in ICC HPV-positive cases (mean)}
  \item{hpv_9v_low}{Relative contribution (\%) of HPV 16/18/31/33/45/52/58 in ICC HPV-positive cases (lower bound of 95\% uncertainty interval)}
  \item{hpv_9v_high}{Relative contribution (\%) of HPV 16/18/31/33/45/52/58 in ICC HPV-positive cases (upper bound of 95\% uncertainty interval)}
  \item{hpv_16}{Relative contribution (\%) of HPV 16 in ICC HPV-positive cases (mean)}
  \item{hpv_16_low}{Relative contribution (\%) of HPV 16 in ICC HPV-positive cases (lower bound of 95\% uncertainty interval)}
  \item{hpv_16_high}{Relative contribution (\%) of HPV 16 in ICC HPV-positive cases (upper bound of 95\% uncertainty interval)}
  \item{hpv_18}{Relative contribution (\%) of HPV 18 in ICC HPV-positive cases (mean)}
  \item{hpv_18_low}{Relative contribution (\%) of HPV 18 in ICC HPV-positive cases (lower bound of 95\% uncertainty interval)}
  \item{hpv_18_high}{Relative contribution (\%) of HPV 18 in ICC HPV-positive cases (upper bound of 95\% uncertainty interval)}
  \item{hpv_31}{Relative contribution (\%) of HPV 31 in ICC HPV-positive cases (mean)}
  \item{hpv_31_low}{Relative contribution (\%) of HPV 31 in ICC HPV-positive cases (lower bound of 95\% uncertainty interval)}
  \item{hpv_31_high}{Relative contribution (\%) of HPV 31 in ICC HPV-positive cases (upper bound of 95\% uncertainty interval)}
  \item{hpv_33}{Relative contribution (\%) of HPV 33 in ICC HPV-positive cases (mean)}
  \item{hpv_33_low}{Relative contribution (\%) of HPV 33 in ICC HPV-positive cases (lower bound of 95\% uncertainty interval)}
  \item{hpv_33_high}{Relative contribution (\%) of HPV 33 in ICC HPV-positive cases (upper bound of 95\% uncertainty interval)}
  \item{hpv_45}{Relative contribution (\%) of HPV 45 in ICC HPV-positive cases (mean)}
  \item{hpv_45_low}{Relative contribution (\%) of HPV 45 in ICC HPV-positive cases (lower bound of 95\% uncertainty interval)}
  \item{hpv_45_high}{Relative contribution (\%) of HPV 45 in ICC HPV-positive cases (upper bound of 95\% uncertainty interval)}
  \item{hpv_52}{Relative contribution (\%) of HPV 52 in ICC HPV-positive cases (mean)}
  \item{hpv_52_low}{Relative contribution (\%) of HPV 52 in ICC HPV-positive cases (lower bound of 95\% uncertainty interval)}
  \item{hpv_52_high}{Relative contribution (\%) of HPV 52 in ICC HPV-positive cases (upper bound of 95\% uncertainty interval)}
  \item{hpv_58}{Relative contribution (\%) of HPV 58 in ICC HPV-positive cases (mean)}
  \item{hpv_58_low}{Relative contribution (\%) of HPV 58 in ICC HPV-positive cases (lower bound of 95\% uncertainty interval)}
  \item{hpv_58_high}{Relative contribution (\%) of HPV 58 in ICC HPV-positive cases (upper bound of 95\% uncertainty interval)}
  \item{hpv_6}{Relative contribution (\%) of HPV 6 in ICC HPV-positive cases (mean)}
  \item{hpv_6_low}{Relative contribution (\%) of HPV 6 in ICC HPV-positive cases (lower bound of 95\% uncertainty interval)}
  \item{hpv_6_high}{Relative contribution (\%) of HPV 6 in ICC HPV-positive cases (upper bound of 95\% uncertainty interval)}
  \item{hpv_11}{Relative contribution (\%) of HPV 11 in ICC HPV-positive cases (mean)}
  \item{hpv_11_low}{Relative contribution (\%) of HPV 11 in ICC HPV-positive cases (lower bound of 95\% uncertainty interval)}
  \item{hpv_11_high}{Relative contribution (\%) of HPV 11 in ICC HPV-positive cases (upper bound of 95\% uncertainty interval)}
}
}
\source{
{Serrano B, Alemany L, Tous S, Bruni L, Clifford GM, Weiss T, et al.
Potential impact of a nine-valent vaccine in human papillomavirus related cervical disease.
Infect Agents Cancer. 2012;7: 38. \url{https://doi.org/10.1186/1750-9378-7-38}}
}
\usage{
data.hpv_distribution
}
\description{
A dataset containing relative contribution of HPV 16/18/31/33/45/52/58 in
cases of ICC HPV-positive, by region and country
}
\keyword{datasets}
