% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.incidence}
\alias{data.incidence}
\title{Incidence of cervical cancer, by age and country}
\format{A data table with 185 observations of 103 variables.
\describe{
  \item{Country}{Country name}
  \item{0..100}{Age 0-100 -- Crude rate, cervix uteri, females, by age; annual rate per individual}
  \item{iso3}{ISO3 country code}
}}
\source{
\url{https://gco.iarc.fr/today/online-analysis-table}
}
\usage{
data.incidence
}
\description{
A dataset containing the incidence of cervical cancer in 185 countries, as reported by
    IARC's Globocan 2018 database. Crude rate, cervix uteri, females, by age.
}
\keyword{datasets}
