% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.mortall}
\alias{data.mortall}
\title{WHO life table}
\format{A data table with 196 observations of 107 variables.
\describe{
  \item{Country ¦ Age [12]}{Country name}
  \item{0..100}{Age 0-100}
  \item{V103..V106}{na}
  \item{iso3}{ISO3 country code}
}}
\source{
\url{}
}
\usage{
data.mortall
}
\description{
A dataset containing the WHO life table.
}
\keyword{datasets}
