% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.valid}
\alias{data.valid}
\title{Global data table}
\format{A data table with 194 observations of 35 variables.
\describe{
  \item{Country}{Country name}
  \item{iso2}{ISO2 country code}
}}
\source{
\url{}
}
\usage{
data.valid
}
\description{
A dataset containing a global range of variables.
}
\keyword{datasets}
