% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{OutputGavi}
\alias{OutputGavi}
\title{Formatting output for VIMC Montagu}
\usage{
OutputGavi(DT, age_stratified = TRUE, calendar_year = FALSE,
  gavi_template = -1)
}
\arguments{
\item{DT}{data table with results}

\item{age_stratified}{logical, whether output should be stratified by age}

\item{calendar_year}{logical, whether output should be given by calendar year
of event OR by year of birth of cohort}

\item{gavi_template}{data table with template file downloaded from montagu}
}
\value{
#
}
\description{
\code{OutputGavi} takes result of BatchRun and outputs it in format to be uploaded to VIMC Montagu.
}
\examples{
#
}
