% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_detect_parallel_plan}
\alias{plc_detect_parallel_plan}
\title{Detect optimal parallel backend (multisession or callr)}
\usage{
plc_detect_parallel_plan(
  num_workers = 2,
  startup_tolerance_sec = 30,
  sleep_sec = 2,
  overlap_threshold = 10,
  runtime_threshold = 25
)
}
\arguments{
\item{num_workers}{Number of workers to test}

\item{startup_tolerance_sec}{Max time allowed
for slow startup without fallback (default 30s)}
}
\value{
One of: "multisession" or "callr"
}
\description{
Allows for long startup (e.g. on slower systems).
Use at setup time to decide plan.
}
