% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_find_bed_files_by_partial_name}
\alias{plc_find_bed_files_by_partial_name}
\title{Find .bed files matching a partial name in a directory}
\usage{
plc_find_bed_files_by_partial_name(dir, partial_name)
}
\value{
A character vector of matching .bed file paths (may be empty).
}
\description{
Find .bed files matching a partial name in a directory
}
