% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_tc_sliding_window}
\alias{plc_tc_sliding_window}
\title{Perform Parallel Sliding Window Expansion on GRanges Tag Clusters}
\usage{
plc_tc_sliding_window(
  granges_obj,
  sld_by = 20,
  ext_dis = 200,
  num_cores = NULL
)
}
\description{
Perform Parallel Sliding Window Expansion on GRanges Tag Clusters
}
