% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{plc_test_scipy_save_npz}
\alias{plc_test_scipy_save_npz}
\title{Internal test to confirm that scipy.sparse.save_npz() works via reticulate.
This ensures compatibility between R, reticulate, scipy, and numpy.}
\usage{
plc_test_scipy_save_npz(log_target = NULL)
}
\description{
Internal test to confirm that scipy.sparse.save_npz() works via reticulate.
This ensures compatibility between R, reticulate, scipy, and numpy.
}
