% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIME.R
\name{PRIME}
\alias{PRIME}
\title{Probabilistic imputation to reduce dropout effects in single cell sequencing data}
\usage{
PRIME(sc_cnt, max_it = 5, alp = 1, err_max = 0.05)
}
\arguments{
\item{sc_cnt}{(M x N) dimensional input matrix. This matrix sould be in count scale (not a log scale)
M: number of genes and N: Number of cells}

\item{max_it}{Maximum number of iteration}

\item{alp}{Shape parameter for a sigmoid function (probabilistic weight)}

\item{err_max}{Error tolerance to stop the iteration}
}
\value{
Normalized imputation result in count scale
}
\description{
Probabilistic imputation to reduce dropout effects in single cell sequencing data
}
\examples{
data(testdata)
PRIME_res <- PRIME(testdata)
pca_res <-prcomp(t(log10(1+PRIME_res)))
plot(pca_res$x[,1], pca_res$x[,2], bg = c("red", "blue","green")[factor(label)],  type = "p", pch = 21, xlab = "PC1", ylab = "PC2")
legend("topleft", title="Cell types", c("G1","S","G2M"), fill = c("red", "blue","green"), horiz = TRUE)

}
\references{
Hyundoo Jeong and Zhandong Liu

 PRIME: a probabilistic imputation method to reduce dropout effects in single cell RNA sequencing
}
\author{
Hyundoo Jeong
}
