% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci.R
\name{PRIMEloci}
\alias{PRIMEloci}
\title{PRIMEloci function}
\usage{
PRIMEloci(
  ctss_rse,
  outdir = "./PRIMEloci_output",
  python_path = "~/.virtualenvs/prime-env",
  score_threshold = 0.75,
  score_diff = 0.1,
  num_cores = NULL,
  keep_tmp = FALSE
)
}
\description{
PRIMEloci function
}
