% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{propSexDebut}
\alias{propSexDebut}
\title{Proportion of girls sexually debuted}
\usage{
propSexDebut(age, country_iso3)
}
\arguments{
\item{age}{age of girls}

\item{country_iso3}{ISO3 country code}
}
\value{
Returns proportion of girls in a given country that has sexually debuted at a given age.
}
\description{
\code{propSexDebut} returns proportion of girls sexually debuted in country \code{country_iso3} at age \code{age}.
}
\examples{
propSexDebut (20, "IND")
propSexDebut (30, "ETH")

}
