% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_functions.r
\name{RegisterBatchDataGavi}
\alias{RegisterBatchDataGavi}
\title{RegisterBatchDataGavi}
\usage{
RegisterBatchDataGavi(gavi_coverage, gavi_template, use_campaigns,
  use_routine, restrict_to_coverage_data = FALSE, force = FALSE,
  psa = 0)
}
\arguments{
\item{force}{}
}
\value{
#
}
\description{
RegisterBatchDataGavi
}
\examples{
#
}
