% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{remove_metadata_and_duplicates}
\alias{remove_metadata_and_duplicates}
\title{Remove Metadata and Duplicate Genomic Ranges}
\usage{
remove_metadata_and_duplicates(gr)
}
\description{
Remove Metadata and Duplicate Genomic Ranges
}
