% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMEloci_utility.R
\name{set_parallel_plan}
\alias{set_parallel_plan}
\title{Set up a robust future parallel plan}
\usage{
set_parallel_plan(num_workers = 1)
}
\description{
Internal utility to set future plan
with callr or multisession fallback.
}
