% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{vectorListToMatrix}
\alias{vectorListToMatrix}
\title{Helper function, export for PRIMEloci
Converts a list of sparse vectors to a sparse matrix.}
\usage{
vectorListToMatrix(vl)
}
\description{
Helper function, export for PRIMEloci
Converts a list of sparse vectors to a sparse matrix.
}
