% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.r
\name{writeBw}
\alias{writeBw}
\title{Export bigWig files for each replicate.}
\usage{
writeBw(object, replicates, dir, inputAssay = "TMM")
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} object containing the CTSSs.}

\item{replicates}{A character vector of replicate names to export.}

\item{dir}{A character string of the directory to export the bigWig files to.}

\item{inputAssay}{A character string of the assay to use for the bigWig scores. Default is "TMM".}
}
\value{
No return value. The function exports bigWig files to the specified directory.
}
\description{
Export bigWig files for each replicate.
}
